Early work on indexer for downloaded files. In downloader when all results have previously been downloaded we now populate the __types.txt file.

+++ src/com/hourglassapps/cpi_ii/web_search/ContentTypeSourceable.java
@@ -1,16 +1,16 @@
 package com.hourglassapps.cpi_ii.web_search;
 
 public class ContentTypeSourceable implements Sourceable {
-	private final long mFilename;
+	private final int mFilename;
 	private final String mContentType;
 	
-	public ContentTypeSourceable(long pFilename, String pContentType) {
+	public ContentTypeSourceable(int pFilename, String pContentType) {
 		mFilename=pFilename;
 		mContentType=pContentType;
 	}
 	
 	@Override
-	public long dstKey() {
+	public int dstKey() {
 		return mFilename;
 	}
