+++ src/com/hourglassapps/cpi_ii/web_search/Sourceable.java
@@ -1,6 +1,6 @@
 package com.hourglassapps.cpi_ii.web_search;
 
 public interface Sourceable {
-	public long dstKey(); 
+	public int dstKey(); 
 	public String src();
 }
