+++ src/com/hourglassapps/util/Downloader.java
@@ -9,5 +9,5 @@ import org.jdeferred.Promise;
 
 public interface Downloader<S,R> {
 	public void reset() throws IOException;
-	public Promise<R,IOException,Void> downloadLink(S pSrc, long pDstKey, Path pDst) throws IOException;
+	public Promise<R,IOException,Void> downloadLink(S pSrc, int pDstKey, Path pDst) throws IOException;
 }
